refactor(theme): wrap the whole Gatsby app with `ThemeProvider`
Gatsby's `wrapRootElement` is perfect for setting up global app providers. The theme provider is used there and we keep control of its behavior by wrapping it up in `App.js` component.

This is particularly useful because a) everything in our app would have access to the theme and b) retain control of this wrapper so we can implement dark-mode and similar features.